Consolidate chip turn completion trigger with onPublicStageDataUpdated
