+++ functions/src/stages/chip.utils.ts
@@ -1,10 +1,17 @@
 import {
   ChipStagePublicData,
+  ChipTransactionStatus,
   ParticipantProfile,
   ParticipantStatus,
+  createChipInfoLogEntry,
+  createChipOfferDeclinedLogEntry,
+  createChipRoundLogEntry,
   createChipTurn,
+  createChipTurnLogEntry,
   sortParticipantsByRandomProfile,
 } from '@deliberation-lab/utils';
+
+import {getFirestoreStagePublicDataRef} from '../utils/firestore';
 
 import * as admin from 'firebase-admin';
 import {Timestamp} from 'firebase-admin/firestore';
@@ -150,4 +157,145 @@ export async function updateParticipantChipQuantities(
   publicStageData.participantChipMap[publicId] = answer.chipMap;
 
   return {answerDoc, answer, publicStageData};
+}
+
+/** Update current turn/round if all participants have responded to
+ * current offer.
+ */
+export async function updateChipTurn(
+  experimentId: string,
+  cohortId: string,
+  stageConfig: ChipStageConfig,
+  publicStage: ChipStagePublicData,
+) {
+  const stageId = stageConfig.id;
+
+  // Define log entry collection reference
+  const logCollection = app
+    .firestore()
+    .collection('experiments')
+    .doc(experimentId)
+    .collection('cohorts')
+    .doc(cohortId)
+    .collection('publicStageData')
+    .doc(stageId)
+    .collection('logs');
+
+  // Define chip transaction collection reference
+  const transactionCollection = app
+    .firestore()
+    .collection('experiments')
+    .doc(experimentId)
+    .collection('cohorts')
+    .doc(cohortId)
+    .collection('publicStageData')
+    .doc(stageId)
+    .collection('transactions');
+
+  await app.firestore().runTransaction(async (transaction) => {
+    const numRounds = stageConfig.numRounds;
+
+    const currentRound = publicStage.currentRound;
+    const currentTurn = publicStage.currentTurn;
+
+    if (!publicStage.participantOfferMap[currentRound]) {
+      return false;
+    }
+
+    const currentTransaction =
+      publicStage.participantOfferMap[currentRound][currentTurn];
+
+    // If no offer, no need to update
+    if (!currentTransaction) {
+      return false;
+    }
+
+    // Check all cohort participants for response to offer
+    const participants = await getChipParticipants(experimentId, cohortId);
+    const participantIds = participants.map((p) => p.publicId);
+
+    const acceptedOffer: string[] = [];
+    for (const participantId of participantIds) {
+      if (
+        participantId !== currentTurn &&
+        !(participantId in currentTransaction.responseMap)
+      ) {
+        // If an active participant (not the current sender) has not
+        // responded. do not proceed
+        return false;
+      } else if (
+        participantId !== currentTurn &&
+        currentTransaction.responseMap[participantId]?.response
+      ) {
+        // Track participants who accepted the current offer
+        acceptedOffer.push(participantId);
+      }
+    }
+
+    const timestamp = Timestamp.now();
+
+    // If all (non-offer) participants have responded to the offer,
+    // execute chip transaction
+    const senderId = currentTurn;
+    const recipientId =
+      acceptedOffer.length > 0
+        ? acceptedOffer[Math.floor(Math.random() * acceptedOffer.length)]
+        : null;
+
+    publicStage.participantOfferMap[currentRound][currentTurn].recipientId =
+      recipientId;
+
+    // Run chip offer transaction and write relevant logs
+    if (recipientId !== null) {
+      currentTransaction.status = ChipTransactionStatus.ACCEPTED;
+      publicStage.participantOfferMap[currentRound][currentTurn] =
+        currentTransaction;
+      // Sender/recipient chips will be updated on chip transaction trigger
+      transaction.set(transactionCollection.doc(), currentTransaction);
+    } else {
+      currentTransaction.status = ChipTransactionStatus.DECLINED;
+      publicStage.participantOfferMap[currentRound][currentTurn] =
+        currentTransaction;
+      transaction.set(
+        logCollection.doc(),
+        createChipOfferDeclinedLogEntry(currentTransaction.offer, timestamp),
+      );
+    }
+
+    // Then, update current turn
+    publicStage.currentTurn = null;
+    const oldCurrentRound = currentRound;
+    const newData = updateChipCurrentTurn(publicStage, participants, numRounds);
+
+    // Write logs
+    if (newData.isGameOver) {
+      transaction.set(
+        logCollection.doc(),
+        createChipInfoLogEntry('The game has ended.', timestamp),
+      );
+    } else {
+      // Write new round log entry if applicable
+      if (oldCurrentRound !== newData.currentRound) {
+        transaction.set(
+          logCollection.doc(),
+          createChipRoundLogEntry(newData.currentRound, timestamp),
+        );
+      }
+      // Write new turn entry
+      transaction.set(
+        logCollection.doc(),
+        createChipTurnLogEntry(
+          newData.currentRound,
+          newData.currentTurn,
+          timestamp,
+        ),
+      );
+    }
+
+    // Update public stage data
+    transaction.set(
+      await getFirestoreStagePublicDataRef(experimentId, cohortId, stageId),
+      newData,
+    );
+  }); // end transaction
 }
